backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
django-cors-headers==4.3.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
celery==5.3.4
django-celery-beat==2.5.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
